+++ src/modules/posts/infrastructure/posts-repositories.ts
@@ -1,12 +1,12 @@
 import { Injectable } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
-import {  Model } from "mongoose";
+import { Model } from "mongoose";
 import { Post, PostDocument } from "../domain/post-schema-Model";
 import { PreparationPostForDB } from "../domain/post-preparation-for-DB";
 import { LikesPostsStatus, LikesPostsStatusDocument } from "../domain/likesPost-schema-Model";
 import { ObjectId } from "mongodb";
-import { CreatePostDto } from "../api/input-Dtos/create-Post-Dto-Model";
 import { PostDBType } from "../domain/post-DB-Type";
+import { CreatePostByBlogIdDto } from "../api/input-Dtos/create-Post-By-BlogId-Dto-Model";
 
 @Injectable()
 export class PostsRepositories {
@@ -18,24 +18,24 @@ export class PostsRepositories {
 
   async createPost(newPost: PreparationPostForDB): Promise<PostDBType> {
     const post = await this.postModel.create(newPost);
-    if (!post) throw new Error('not today server all (')
-    return post
+    if (!post) throw new Error("not today server all (");
+    return post;
   }
 
-  async deletePost(id: string): Promise<boolean> {
+  async deletePost(id: string, userId: string): Promise<boolean> {
     const result = await this.postModel
-      .deleteOne({ _id: new ObjectId(id) })
+      .deleteOne({ _id: new ObjectId(id), userId: userId })
       .exec();
     return result.deletedCount === 1;
   }
 
-  async updatePost(id: string, data: CreatePostDto): Promise<boolean> {
-    const result = await this.postModel.updateOne({ _id: new ObjectId(id) }, {
+  async updatePost(id: string, data: CreatePostByBlogIdDto, blogId: string, userId: string): Promise<boolean> {
+    const result = await this.postModel.updateOne({ _id: new ObjectId(id), userId: userId }, {
       $set: {
         title: data.title,
         shortDescription: data.shortDescription,
         content: data.content,
-        blogId: data.blogId
+        blogId: blogId
       }
     });
     return result.matchedCount === 1;
@@ -43,7 +43,7 @@ export class PostsRepositories {
 
   async findPost(id: string): Promise<PostDBType> {
     const post = await this.postModel.findOne({ _id: new ObjectId(id) });
-    if(!post) return null
+    if (!post) return null;
     return post;
   }
 
@@ -52,7 +52,7 @@ export class PostsRepositories {
       { userId: userId, parentId: id },
       { $set: { likeStatus: likeStatus, addedAt: new Date().toISOString(), login: login } },
       { upsert: true });
-    if (!like) return null
+    if (!like) return null;
     return true;
   }
 
@@ -78,4 +78,9 @@ export class PostsRepositories {
       newComment.createdAt,
       likesInfo)
   }*/
+  async updateStatusBan(userId: string, isBanned: boolean): Promise<boolean> {
+    const result = await this.postModel.updateMany({ userId: userId },
+      { $set: { isBanned: isBanned }});
+    return result.matchedCount === 1;
+  }
 }
